Add files via upload
this is similar to my previous code, just that it prompts you to chose the file to open. then you just need to call the function plot_temperature_graph() in the console, to get the plot to show. I found it easier to do that, and then decide when i want to plot instead of always plotting as it can be sometimes slow.
will add for RH soon.